Implement new strategy for tracking transaction depth changes related to errors
Instead of providing API for custom error handling on commit, connection is free to register updates to the transaction manager status as it encounters errors.

This transaction manager status is updated to be able to store that transaction is aborted and requires top-level rollback to be usable again.
This information is used in `AnsiTransactionManager`'s commit/rollback logic.